+++ src/components/vulnerability-db/ArticleDetailHeader.tsx
@@ -1,0 +1,35 @@
+import { formatTimestampAsDateTime } from "@/lib/utils";
+import { VulDBPost } from "@/types/post";
+import { Timestamp } from "firebase/firestore";
+import { IconPin, IconShare } from "../ui/Icons";
+
+export function ArticleDetailHeader({ post }: { post: VulDBPost }) {
+  const haveUpdatedAt = post.source_updated_at;
+
+  return (
+    <section className="w-[82.125rem] border-b border-b-line-default p-[1.75rem_0_3.75rem_0]">
+      <h1 className="mb-8 mt-[1.261rem] text-4xl font-medium leading-[2.723rem] tracking-[-0.01em]">
+        {post?.title?.translated || "제목이 없습니다."}
+      </h1>
+      <div className="flex items-center justify-between text-xl font-normal leading-[1.512rem] text-gray-default">
+        <div className="flex gap-9 tracking-[-0.01em]">
+          <p>취약성 뉴스 세부정보</p>
+          <p>
+            업로드시간 |{" "}
+            {post.created_at ? formatTimestampAsDateTime(post.created_at) : ""}
+          </p>
+          <p>
+            출시시간 |{" "}
+            {haveUpdatedAt
+              ? formatTimestampAsDateTime(post.source_updated_at as Timestamp)
+              : formatTimestampAsDateTime(post.source_created_at)}
+          </p>
+        </div>
+        <div className="flex gap-[1.625rem]">
+          <IconPin />
+          <IconShare />
+        </div>
+      </div>
+    </section>
+  );
+}
